Engine: moved dirty rect code into separate file
